Tipcalculator
This Python based tip calculator takes the guesswork out of splitting a bill! Just enter the total bill (in dollars),the percentage of tip you'd like to give, and the number of people sharing. The program the calculates each person's fair share, making group dining easier and ensuring everyone contributes accurately. Perfect for a smooth and enjoyable meal!